+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unpublished]
 
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+## [0.11.0] 2023-02-02
+
 Breaking changes
 
 ### Added